+++ .github/workflows/test_common.yml
@@ -18,8 +18,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
+        python-version: [3.9]
         runs-on: [ubuntu-22.04, windows-2019, macos-13]
-        python-version: ["3.9"]
 
     runs-on: ${{ matrix.runs-on }}
 
@@ -35,6 +35,7 @@ jobs:
       - name: Install dependencies
         run: |
           pip install --upgrade pip
+          pip install --no-cache-dir torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
           pip install .[tests]
 
       - name: Test with pytest
