+++ .github/workflows/test_export_tflite_cli.yml
@@ -5,11 +5,7 @@ on:
     branches: [main]
   pull_request:
     branches: [main]
-    paths:
-      - ".github/workflows/test_export_cli.yml"
-      - "optimum/exporters/tflite/**.py"
-      - "tests/exporters/tflite/**.py"
-      - "setup.py"
+    types: [opened, synchronize, reopened, labeled, unlabeled]
 
 concurrency:
   group: ${{ github.workflow }}-${{ github.head_ref || github.run_id }}
@@ -20,6 +16,8 @@ env:
 
 jobs:
   build:
+    if: ${{ (github.event_name == 'push') || contains( github.event.pull_request.labels.*.name, 'tflite' ) }}
+
     strategy:
       fail-fast: false
       matrix:
