+++ tests/exporters/onnx/test_export.py
@@ -14,6 +14,8 @@
 # limitations under the License.
 import gc
 import os
+import tempfile
+import unittest
 from functools import partial
 from pathlib import Path
 from tempfile import TemporaryDirectory
@@ -21,15 +23,27 @@ from typing import Dict
 from unittest import TestCase, mock
 
 import onnx
+import onnxruntime
 import pytest
+import torch
 from parameterized import parameterized
-from transformers import AutoConfig, is_tf_available, is_torch_available
+from transformers import (
+    AutoConfig,
+    AutoModelForSeq2SeqLM,
+    AutoModelForSequenceClassification,
+    is_tf_available,
+    is_torch_available,
+)
+from transformers.modeling_utils import PreTrainedModel
 from transformers.testing_utils import require_onnx, require_tf, require_torch, require_torch_gpu, require_vision, slow
 
+from optimum.exporters import TasksManager
 from optimum.exporters.error_utils import AtolError
 from optimum.exporters.onnx import (
     OnnxConfig,
+    OnnxConfigWithLoss,
     OnnxConfigWithPast,
+    export,
     export_models,
     get_decoder_models_for_export,
     get_diffusion_models_for_export,
@@ -43,11 +57,13 @@ from optimum.exporters.onnx.config import TextDecoderOnnxConfig
 from optimum.exporters.onnx.constants import SDPA_ARCHS_ONNX_EXPORT_NOT_SUPPORTED
 from optimum.exporters.onnx.model_configs import WhisperOnnxConfig
 from optimum.exporters.onnx.utils import get_speecht5_models_for_export
-from optimum.utils import DummyPastKeyValuesGenerator, NormalizedTextConfig
+from optimum.onnxruntime.utils import ONNX_DECODER_NAME
+from optimum.utils import DummyPastKeyValuesGenerator, DummyTextInputGenerator, NormalizedTextConfig
+from optimum.utils.normalized_config import NormalizedConfigManager
 from optimum.utils.save_utils import maybe_load_preprocessors
 from optimum.utils.testing_utils import grid_parameters, require_diffusers
 
-from ..exporters_utils import (
+from ..utils import (
     PYTORCH_DIFFUSION_MODEL,
     PYTORCH_EXPORT_MODELS_TINY,
     PYTORCH_SENTENCE_TRANSFORMERS_MODEL,
@@ -57,9 +73,9 @@ from ..exporters_utils import (
 )
 
 
-if is_torch_available() or is_tf_available():
-    from optimum.exporters.tasks import TasksManager
-
+if is_tf_available():
+    from transformers import TFAutoModelForSequenceClassification
+    from transformers.modeling_tf_utils import TFPreTrainedModel
 
 SEED = 42
 
@@ -688,3 +704,199 @@ class OnnxExportModelTest(TestCase):
             monolith,
             device="cuda",
         )
+
+
+class TestOnnxConfigWithLoss(unittest.TestCase):
+    @pytest.mark.tensorflow_test
+    def test_onnx_config_with_loss(self):
+        # Prepare model and dataset
+        model_checkpoint = "hf-internal-testing/tiny-random-bert"
+        models = {
+            AutoModelForSequenceClassification.from_pretrained(model_checkpoint),
+        }
+        if is_tf_available():
+            models.add(TFAutoModelForSequenceClassification.from_pretrained(model_checkpoint))
+
+        for model in models:
+            with self.subTest(model=model):
+                with tempfile.TemporaryDirectory() as tmp_dir:
+                    onnx_config_constructor = TasksManager.get_exporter_config_constructor(
+                        model=model, exporter="onnx", task="text-classification"
+                    )
+                    onnx_config = onnx_config_constructor(model.config)
+
+                    wrapped_onnx_config = OnnxConfigWithLoss(onnx_config)
+
+                    # Export model from PyTorch to ONNX
+                    onnx_model_path = Path(os.path.join(tmp_dir, f"{model.config.model_type}.onnx"))
+                    opset = max(onnx_config.DEFAULT_ONNX_OPSET, 12)
+                    _ = export(
+                        model=model,
+                        config=wrapped_onnx_config,
+                        opset=opset,
+                        output=onnx_model_path,
+                    )
+
+                    # ONNX Runtime Inference
+                    ort_sess = onnxruntime.InferenceSession(
+                        onnx_model_path.as_posix(),
+                        providers=[
+                            (
+                                "CUDAExecutionProvider"
+                                if torch.cuda.is_available()
+                                and "CUDAExecutionProvider" in onnxruntime.get_available_providers()
+                                else "CPUExecutionProvider"
+                            )
+                        ],
+                    )
+                    framework = "pt" if isinstance(model, PreTrainedModel) else "tf"
+                    normalized_config = NormalizedConfigManager.get_normalized_config_class("bert")(model.config)
+                    input_generator = DummyTextInputGenerator(
+                        "text-classification", normalized_config, batch_size=2, sequence_length=16
+                    )
+
+                    inputs = {
+                        name: input_generator.generate(name, framework=framework)
+                        for name in ["input_ids", "attention_mask", "token_type_ids"]
+                    }
+                    inputs["labels"] = input_generator.constant_tensor(
+                        [2], value=0, dtype=inputs["input_ids"].dtype, framework=framework
+                    )
+
+                    input_names = [ort_input.name for ort_input in ort_sess._inputs_meta]
+                    output_names = [output.name for output in ort_sess._outputs_meta]
+
+                    input_feed = {input_name: inputs[input_name].cpu().numpy() for input_name in input_names}
+
+                    ort_outputs = ort_sess.run(output_names, input_feed)
+                    pt_outputs = model(**inputs)
+
+                    # Checkers
+                    assert len(ort_outputs) > 1, "There is only one element in outputs, the loss might be missing!"
+                    if issubclass(type(model), PreTrainedModel):
+                        self.assertAlmostEqual(
+                            float(ort_outputs[0]),
+                            float(pt_outputs["loss"]),
+                            3,
+                            "The losses of ONNX Runtime and PyTorch inference are not close enough!",
+                        )
+                    elif issubclass(type(model), TFPreTrainedModel):
+                        for ort_loss, pt_loss in zip(ort_outputs[-1], pt_outputs["loss"]):
+                            self.assertAlmostEqual(
+                                float(ort_loss),
+                                float(pt_loss),
+                                3,
+                                "The losses of ONNX Runtime and PyTorch inference are not close enough!",
+                            )
+                    gc.collect()
+
+    def test_onnx_decoder_model_with_config_with_loss(self):
+        with tempfile.TemporaryDirectory() as tmp_dir:
+            # Prepare model and dataset
+            model_checkpoint = "hf-internal-testing/tiny-random-gpt2"
+            model = AutoModelForSequenceClassification.from_pretrained(model_checkpoint)
+
+            # Wrap OnnxConfig
+            onnx_config_constructor = TasksManager.get_exporter_config_constructor(
+                model=model, exporter="onnx", task="text-classification"
+            )
+            onnx_config = onnx_config_constructor(model.config)
+            wrapped_onnx_config = OnnxConfigWithLoss(onnx_config)
+
+            # Export model from PyTorch to ONNX
+            onnx_model_path = Path(os.path.join(tmp_dir, f"{model.config.model_type}.onnx"))
+            opset = max(onnx_config.DEFAULT_ONNX_OPSET, 12)
+            export(model=model, config=wrapped_onnx_config, opset=opset, output=onnx_model_path)
+
+            # ONNX Runtime Inference
+            ort_sess = onnxruntime.InferenceSession(
+                onnx_model_path.as_posix(),
+                providers=[
+                    (
+                        "CUDAExecutionProvider"
+                        if torch.cuda.is_available()
+                        and "CUDAExecutionProvider" in onnxruntime.get_available_providers()
+                        else "CPUExecutionProvider"
+                    )
+                ],
+            )
+
+            batch = 3
+            seq_length = 8
+            inputs = {
+                "input_ids": torch.ones((batch, seq_length), dtype=torch.long),
+                "attention_mask": torch.ones((batch, seq_length), dtype=torch.long),
+                "labels": torch.zeros(batch, dtype=torch.long),
+            }
+            input_names = [ort_input.name for ort_input in ort_sess._inputs_meta]
+            output_names = [output.name for output in ort_sess._outputs_meta]
+            input_feed = {input_name: inputs[input_name].cpu().numpy() for input_name in input_names}
+            ort_outputs = ort_sess.run(output_names, input_feed)
+            pt_outputs = model(**inputs)
+
+            # Checkers
+            assert len(ort_outputs) > 1, "There is only one element in outputs, the loss might be missing!"
+            self.assertAlmostEqual(
+                float(ort_outputs[0]),
+                float(pt_outputs["loss"]),
+                3,
+                "The losses of ONNX Runtime and PyTorch inference are not close enough!",
+            )
+            gc.collect()
+
+    def test_onnx_seq2seq_model_with_config_with_loss(self):
+        with tempfile.TemporaryDirectory() as tmp_dir:
+            # Prepare model and dataset
+            model_checkpoint = "hf-internal-testing/tiny-random-t5"
+            model = AutoModelForSeq2SeqLM.from_pretrained(model_checkpoint)
+
+            # Wrap OnnxConfig(decoders)
+            onnx_config_constructor = TasksManager.get_exporter_config_constructor(
+                model=model, exporter="onnx", task="text2text-generation"
+            )
+            onnx_config = onnx_config_constructor(model.config)
+
+            onnx_config_decoder = onnx_config.with_behavior("decoder", use_past=False)
+
+            wrapped_onnx_config_decoder = OnnxConfigWithLoss(onnx_config_decoder)
+
+            # Export decoder models from PyTorch to ONNX
+            opset = max(onnx_config.DEFAULT_ONNX_OPSET, 12)
+
+            onnx_model_path = Path(tmp_dir).joinpath(ONNX_DECODER_NAME)
+
+            export(
+                model=model,
+                config=wrapped_onnx_config_decoder,
+                opset=opset,
+                output=onnx_model_path,
+            )
+
+            # ONNX Runtime Inference
+            ort_sess = onnxruntime.InferenceSession(
+                onnx_model_path.as_posix(),
+                providers=[
+                    (
+                        "CUDAExecutionProvider"
+                        if torch.cuda.is_available()
+                        and "CUDAExecutionProvider" in onnxruntime.get_available_providers()
+                        else "CPUExecutionProvider"
+                    )
+                ],
+            )
+            batch = 3
+            encoder_seq_length = 8
+            encoder_hidden_states_shape = (batch, encoder_seq_length, model.config.hidden_size)
+            inputs = {
+                "input_ids": torch.ones((batch, encoder_seq_length), dtype=torch.long),
+                "encoder_hidden_states": torch.zeros(encoder_hidden_states_shape),
+                "encoder_attention_mask": torch.ones((batch, encoder_seq_length), dtype=torch.long),
+                "labels": torch.zeros((batch, encoder_seq_length), dtype=torch.long),
+            }
+            input_names = [ort_input.name for ort_input in ort_sess._inputs_meta]
+            output_names = [output.name for output in ort_sess._outputs_meta]
+            input_feed = {input_name: inputs[input_name].cpu().numpy() for input_name in input_names}
+
+            ort_sess.run(output_names, input_feed)
+
+            gc.collect()
