+++ tests/exporters/onnx/test_export_cli.py
@@ -39,7 +39,7 @@ from optimum.utils.testing_utils import grid_parameters, require_diffusers, requ
 if is_torch_available():
     from optimum.exporters.tasks import TasksManager
 
-from ..exporters_utils import (
+from ..utils import (
     NO_DYNAMIC_AXES_EXPORT_SHAPES_TRANSFORMERS,
     PYTORCH_DIFFUSION_MODEL,
     PYTORCH_EXPORT_MODELS_TINY,
