+++ tests/exporters/tflite/test_export_cli.py
@@ -24,7 +24,7 @@ from transformers.testing_utils import require_tf
 from optimum.utils import DEFAULT_DUMMY_SHAPES
 
 from ...utils.test_task_processors import TASK_TO_NON_DEFAULT_DATASET
-from ..exporters_utils import PYTORCH_EXPORT_MODELS_TINY
+from ..utils import PYTORCH_EXPORT_MODELS_TINY
 
 
 if is_tf_available():
